Force mutable Strings to address upcoming Ruby default frozen strings.
Ruby is slowly moving towards using frozen string literals by default [1]. As
of Ruby 3.4 it now produces a warning whenever a String literal modifies a
value which will be default frozen in the future.

From this project:

```
lib/graphviz.rb:533: warning: literal string will be frozen in the future (run with --debug-frozen-string-literal for more information)
lib/graphviz/dot_script.rb:58: warning: literal string will be frozen in the future (run with --debug-frozen-string-literal for more information)
```

This change minimally modifies the places where a mutable String is required
to use String.new instead of a bare String literal. This is a no-op for
existing apps not using default frozen strings, but allows all the benefits
for applications supporting it.

Running the tests with the upcoming behavior enabled on Ruby 3.4:

```
RUBYOPT="--enable=frozen_string_literal" rake
```

Before:

```
100 tests, 303 assertions, 71 failures, 5 errors, 0 pendings, 0 omissions, 0 notifications
24% passed
```

After:

```
100 tests, 258 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
```

[1] https://bugs.ruby-lang.org/issues/20205